+++ go/chisel/share/tunnel/tunnel_radius.go
@@ -14,7 +14,7 @@ import (
 	"time"
 
 	"github.com/inverse-inc/go-utils/sharedutils"
-	"github.com/inverse-inc/packetfence/go/chisel/share/tunnel/radius_proxy"
+	"github.com/inverse-inc/packetfence/go/chisel/share/radius_proxy"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
